% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{dpl}
\alias{dpl}
\title{Density, distribution function, quantile function and random generation for the power logistic distribution with parameters mu and sigma.}
\usage{
dpl(x, lambda, mu = 0, sigma = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation for the power logistic distribution with parameters mu and sigma.
}
\examples{
dpl(1, 1, 3, 4)
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.
}
