% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{drpc}
\alias{drpc}
\title{Reversal power cauchy distribution}
\usage{
drpc(x, lambda, mu = 0, sigma = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}
}
\description{
Density function for the reversal power cauchy distribution with parameters mu, sigma and lambda.
}
\examples{
drpc(1, 1, 3, 4)
}
\references{
Bazán, J. L., Torres‐Avilés, F., Suzuki, A. K., & Louzada, F. (2017). Power and reversal power links for binary regressions: An application for motor insurance policyholders. Applied Stochastic Models in Business and Industry, 33(1), 22-34.
}
