% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpn.R, R/ppn.R, R/qpn.R, R/rpn.R
\name{pn}
\alias{pn}
\alias{dpn}
\alias{pn}
\alias{ppn}
\alias{pn}
\alias{qpn}
\alias{pn}
\alias{rpn}
\title{The Power Normal Distribution}
\usage{
dpn(x, lambda, mu = 0, sigma = 1)

ppn(q, lambda, mu = 0, sigma = 1)

qpn(p, lambda, mu = 0, sigma = 1)

rpn(n, lambda, mu = 0, sigma = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function,
quantile function and random generation for
the power normal distribution with parameters mu, sigma and lambda.
}
\examples{
dpn(1, 1, 3, 4)
ppn(1, 1, 3, 4)
qpn(0.2, 1, 3, 4)
rpl(5, 2, 3, 4)
}
\references{
Bazán, J. L., Torres‐Avilés, F., Suzuki, A. K., & Louzada, F. (2017). Power and reversal power links for binary regressions: An application for motor insurance policyholders. Applied Stochastic Models in Business and Industry, 33(1), 22-34.
}
