% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powdist.R
\docType{package}
\name{powdist-package}
\alias{powdist-package}
\alias{powdist}
\title{Power and reversal power distributions}
\description{
The \pkg{powdist} package enables to compute the density,
distribution function, quantile function and random generation
for the family of power and reversal power distributions.
}
\author{
Susan A. C. Anyosa <susanch@usp.br>
}
