% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powdist-pl.R
\docType{package}
\name{powdist-pl}
\alias{powdist-pl}
\alias{ppl}
\alias{powdist-pl-package}
\title{Power logistic distribution}
\usage{
dpl(x, lambda, mu = 0, sigma = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}
}
\description{
Density, distribution function, quantile function and random generation
for the power logistic distribution with parameters mu, sigma and lambda.
}
\examples{
dpl(1, 1, 3, 4)
ppl(1, 1, 3, 4)
qpl(0.2, 1, 3, 4)
rpl(5, 2, 3, 4)
}
\references{
Bazán, J. L., Torres‐Avilés, F., Suzuki, A. K., & Louzada, F. (2017). Power and reversal power links for binary regressions: An application for motor insurance policyholders. Applied Stochastic Models in Business and Industry, 33(1), 22-34.
}
