% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{ppl}
\alias{ppl}
\title{Power logistic distribution}
\usage{
ppl(q, lambda, mu = 0, sigma = 1)
}
\arguments{
\item{q}{vector of quantiles.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}
}
\description{
Power logistic distribution
}
\examples{
ppl(1, 1, 3, 4)
}
\references{
Bazán, J. L., Torres‐Avilés, F., Suzuki, A. K., & Louzada, F. (2017). Power and reversal power links for binary regressions: An application for motor insurance policyholders. Applied Stochastic Models in Business and Industry, 33(1), 22-34.
}
