% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pn.R
\name{ppn}
\alias{ppn}
\title{Power normal distribution}
\usage{
ppn(q, lambda, mu = 0, sigma = 1)
}
\arguments{
\item{q}{vector of quantiles.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}
}
\description{
Distribution function for the power normal distribution with parameters mu, sigma and lambda.
}
\examples{
ppn(1, 1, 3, 4)
}
\references{
Bazán, J. L., Torres‐Avilés, F., Suzuki, A. K., & Louzada, F. (2017). Power and reversal power links for binary regressions: An application for motor insurance policyholders. Applied Stochastic Models in Business and Industry, 33(1), 22-34.
}
