% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drpnorm.R, R/prpnorm.R, R/qrpnorm.R,
%   R/rrpnorm.R
\name{ReversalPowerNormal}
\alias{ReversalPowerNormal}
\alias{drpnorm}
\alias{ReversalPowerNormal}
\alias{prpnorm}
\alias{ReversalPowerNormal}
\alias{qrpnorm}
\alias{ReversalPowerNormal}
\alias{rrpnorm}
\title{The Reversal Power Normal Distribution}
\usage{
drpnorm(x, lambda = 1, mu = 0, sigma = 1, log = FALSE)

prpnorm(q, lambda = 1, mu = 0, sigma = 1, lower.tail = TRUE,
  log.p = FALSE)

qrpnorm(p, lambda = 1, mu = 0, sigma = 1, lower.tail = TRUE,
  log.p = FALSE)

rrpnorm(n, lambda = 1, mu = 0, sigma = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x ]}, otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function,
quantile function and random generation for
the reversal power normal distribution with parameters mu, sigma and lambda.
}
\details{
The reversal power Normal distribution has density

\eqn{f(x)=\lambda \left [ \Phi \left ( -\frac{x-\mu}{\sigma} \right ) \right ]^{\lambda - 1} \left[\frac{e^{ -\frac{1}{2}\left ( \frac{x-\mu}{\sigma} \right )^2}}{\sigma\sqrt{2\pi}} \right]}{f(x)=[\lambda/\sigma][exp(-((x-\mu)/\sigma)^2)/\sqrt(2\pi)][\Phi(-(x-\mu)/\sigma)]^(\lambda-1)},

where \eqn{-\infty<\mu<\infty} is the location paramether, \eqn{\sigma^2>0} the scale parameter and \eqn{\lambda>0} the shape parameter.
}
\examples{
drpnorm(1, 1, 3, 4)
prpnorm(1, 1, 3, 4)
qrpnorm(0.2, 1, 3, 4)
rrpnorm(5, 2, 3, 4)
}
\references{
Anyosa, S. A. C. (2017) \emph{Binary regression using power and reversal power links}. Master's thesis in Portuguese. Interinstitutional Graduate Program in Statistics. Universidade de São Paulo - Universidade Federal de São Carlos. Available in \url{https://repositorio.ufscar.br/handle/ufscar/9016}.

Bazán, J. L., Torres -Avilés, F., Suzuki, A. K. and Louzada, F. (2017) Power and reversal power links for binary regressions: An application for motor insurance policyholders. \emph{Applied Stochastic Models in Business and Industry}, \strong{33}(1), 22-34.

Bazán, J. L., Romeo, J. S. and Rodrigues, J. (2014) Bayesian skew-probit regression for binary response data. \emph{Brazilian Journal of Probability and Statistics}. \strong{28}(4), 467–482.
}
