% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drpnormp.R, R/prpnormp.R, R/qrpnormp.R,
%   R/rrpnormp.R
\name{ReversalPowerPowerExponential}
\alias{ReversalPowerPowerExponential}
\alias{drpnormp}
\alias{ReversalPowerPowerExponential}
\alias{prpnormp}
\alias{ReversalPowerPowerExponential}
\alias{qrpnormp}
\alias{ReversalPowerPowerExponential}
\alias{rrpnormp}
\title{The Reversal Power Power Exponential Distribution}
\usage{
drpnormp(x, lambda = 1, mu = 0, sigma = 1, k = 0, log = FALSE)

prpnormp(q, lambda = 1, mu = 0, sigma = 1, k = 0, lower.tail = TRUE,
  log.p = FALSE)

qrpnormp(p, lambda = 1, mu = 0, sigma = 1, k = 0, lower.tail = TRUE,
  log.p = FALSE)

rrpnormp(n, lambda = 1, mu = 0, sigma = 1, k = 0)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu, sigma}{location and scale parameters.}

\item{k, lambda}{shape parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x ]}, otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function,
quantile function and random generation for
the reversal power power exponential distribution with parameters mu, sigma, lambda and k.
}
\details{
The reversal power power Exponential distribution has density

\eqn{f(x)=[\lambda/\sigma][f((x-\mu)/\sigma)][F((x-\mu)/\sigma)] ^(\lambda-1)},

where \eqn{-\infty<\mu<\infty} is the location paramether, \eqn{\sigma^2>0} the scale parameter and \eqn{\lambda>0} the shape parameter.
}
\examples{
drpnormp(1, 1, 3, 4, 1)
prpnormp(1, 1, 3, 4, 1)
qrpnormp(0.2, 1, 3, 4, 1)
rrpnormp(5, 2, 3, 4, 1)
}
