% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drpl.R, R/prpl.R, R/qrpl.R, R/rrpl.R
\name{revpl}
\alias{revpl}
\alias{drpl}
\alias{revpl}
\alias{prpl}
\alias{revpl}
\alias{qrpl}
\alias{revpl}
\alias{rrpl}
\title{The Reversal Power Logistic Distribution}
\usage{
drpl(x, lambda, mu = 0, sigma = 1)

prpl(q, lambda, mu = 0, sigma = 1)

qrpl(p, lambda, mu = 0, sigma = 1)

rrpl(n, lambda, mu = 0, sigma = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function,
quantile function and random generation for
the reversal power logistic distribution with parameters mu, sigma and lambda.
}
\examples{
drpl(1, 1, 3, 4)
prpl(1, 1, 3, 4)
qrpl(0.2, 1, 3, 4)
rrpl(5, 2, 3, 4)
}
\references{
Bazán, J. L., Torres‐Avilés, F., Suzuki, A. K., & Louzada, F. (2017). Power and reversal power links for binary regressions: An application for motor insurance policyholders. Applied Stochastic Models in Business and Industry, 33(1), 22-34.
}
