% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpgmiv.R
\name{rrpgmiv}
\alias{rrpgmiv}
\title{Reversal power gumbel of minimum value distribution}
\usage{
rrpgmiv(n, lambda, mu = 0, sigma = 1)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}
}
\description{
Random generation for the reversal power gumbel of minimum value distribution with parameters mu, sigma and lambda.
}
\examples{
rrpgmiv(5, 2, 3, 4)
}
\references{
Bazán, J. L., Torres‐Avilés, F., Suzuki, A. K., & Louzada, F. (2017). Power and reversal power links for binary regressions: An application for motor insurance policyholders. Applied Stochastic Models in Business and Industry, 33(1), 22-34.
}
