% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgmiv.R, R/ppgmiv.R, R/qpgmiv.R, R/rpgmiv.R
\name{pgmiv}
\alias{pgmiv}
\alias{dpgmiv}
\alias{pgmiv}
\alias{ppgmiv}
\alias{pgmiv}
\alias{qpgmiv}
\alias{pgmiv}
\alias{rpgmiv}
\title{The Power Gumbel of Minimum Value Distribution}
\usage{
dpgmiv(x, lambda, mu = 0, sigma = 1)

ppgmiv(q, lambda, mu = 0, sigma = 1)

qpgmiv(p, lambda, mu = 0, sigma = 1)

rpgmiv(n, lambda, mu = 0, sigma = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function,
quantile function and random generation for
the power Gumbel of minimum value distribution with parameters mu, sigma and lambda.
}
\examples{
dpgmiv(1, 1, 3, 4)
ppgmiv(1, 1, 3, 4)
qpgmiv(0.2, 1, 3, 4)
rpgmiv(5, 2, 3, 4)
}
\references{
Abanto -Valle, C. A., Bazán, J. L. and Smith, A. C. (2014) \emph{State space mixed models for binary responses with skewed inverse links using JAGS}. Rio de Janeiro, Brazil.

Anyosa, S. A. C. (2017) \emph{Binary regression using power and reversal power links}. Master's thesis in Portuguese. Interinstitutional Graduate Program in Statistics. Universidade de São Paulo - Universidade Federal de São Carlos. Available in \url{http://conteudo.icmc.usp.br/pessoas/jlbazan/download/thesis.pdf}.

Bazán, J. L., Torres -Avilés, F., Suzuki, A. K. and Louzada, F. (2017) Power and reversal power links for binary regressions: An application for motor insurance policyholders. \emph{Applied Stochastic Models in Business and Industry}, \strong{33}(1), 22-34.
}
