% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powdist-package.R, R/powdist-pl.R
\docType{package}
\name{powdist-package}
\alias{powdist-package}
\alias{powdist}
\alias{powdist-package}
\alias{ppl}
\title{Power and reversal power distributions}
\usage{
dnorm(x, mean = 0, sd = 1, log = FALSE)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}
}
\description{
The \pkg{powdist} package enables to compute the probability density function,
cumulative distribution function, quantile function and generate random numbers for
the following distributions: power logistic (pl), reversal power logistic (rpl),
power normal (pn), reversal power normal (rpn), power cauchy (pc), reversal power cauchy (rpc),
power gumbel of maximum value (pgmav), reversal power gumbel of maximum value (rpgmav),
power gumbel of minimum value (pgmiv) and reversal power gumbel of minimum value (rpgmiv).

Density, distribution function, quantile function and random generation
for the normal distribution with mean equal to mean and standard deviation
equal to sd.
}
\examples{
dpl(1, 1, 3, 4)
}
\references{
Bazán, J. L., Torres‐Avilés, F., Suzuki, A. K., & Louzada, F. (2017). Power and reversal power links for binary regressions: An application for motor insurance policyholders. Applied Stochastic Models in Business and Industry, 33(1), 22-34.
}
