% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powdist-package.R
\docType{package}
\name{powdist-package}
\alias{powdist-package}
\alias{powdist}
\title{Power and reversal power distributions}
\description{
The \pkg{powdist} package enables to compute the probability density function,
cumulative distribution function, quantile function and generate random numbers for
the following distributions: power logistic, reversal power logistic, power normal,
reversal power normal, power cauchy, reversal power cauchy, power gumbel of maximum value,
reversal power gumbel of maximum value, power gumbel of minimum value and reversal
power gumbel of maximum value.
}
