% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powdist-package.R
\docType{package}
\name{powdist-package}
\alias{powdist-package}
\alias{powdist}
\title{Power and reversal power distributions}
\description{
The \pkg{powdist} package enables to compute the probability density function,
cumulative distribution function, quantile function and generate random numbers for
the following distributions: power Logistic (plogis), reversal power Logistic (rplogis),
power Normal (pnorm), reversal power Normal (rpnorm), power Cauchy (pcauchy), reversal power Cauchy (rpcauchy),
power reversal-Gumbel (prgumbel),
power Student T (pt), reversal power Student T (rpt), power Laplace (plaplace),
reversal power Laplace (rplaplace), power exponential power (pexpow) and
reversal power exponential power (rpexpow).
}
