% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powdist.R
\name{powdist}
\alias{powdist}
\title{Density, distribution function, quantile function and random generation for the family of power and reversal power distributions..}
\description{
Density, distribution function, quantile function and random generation for the family of power and reversal power distributions..
}
\author{
Susan A. C. Anyosa <susanch@usp.br>
}
