% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{ppl}
\alias{ppl}
\title{Power logistic distribution}
\usage{
ppl(q, lambda, mu = 0, sigma = 1)
}
\arguments{
\item{lambda}{shape parameter.}

\item{mu, sigma}{location and scale parameters.}

\item{x, q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation for the normal distribution with mean equal to mean and standard deviation equal to sd.
}
\examples{
dpl(1, 1, 3, 4)
}
\references{
Bazán, J. L., Torres‐Avilés, F., Suzuki, A. K., & Louzada, F. (2017). Power and reversal power links for binary regressions: An application for motor insurance policyholders. Applied Stochastic Models in Business and Industry, 33(1), 22-34.
}
